# artifacts/api-server/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Order Automation System API",
    "description": "API for automated payment and order management system. Includes product management, order processing, webhook handling, and admin reporting.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development server"
    },
    {
      "url": "https://order-automation-api.onrender.com/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/webhook/sepay": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "SePay webhook",
        "description": "Process payment notifications from SePay",
        "operationId": "sepayWebhook",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SePayWebhookPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/webhook/retry-status": {
      "get": {
        "tags": ["Webhooks"],
        "summary": "Get webhook retry queue status",
        "description": "Admin endpoint to check webhook retry queue status",
        "operationId": "getWebhookRetryStatus",
        "security": [
          {
            "X-Telegram-Chat-Id": []
          }
        ],
        "responses": {
          "200": {
            "description": "Retry queue status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookRetryStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "SePay webhook token (format: Apikey <token>)"
      },
      "X-CSRF-Token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF protection token"
      },
      "X-Telegram-Chat-Id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Telegram-Chat-Id",
        "description": "Telegram admin chat ID for authentication"
      }
    },
    "schemas": {
...
      "SePayWebhookPayload": {
        "type": "object",
        "required": ["id", "bankBrandName", "accountNumber", "transactionDate", "amountIn", "amountOut", "content", "transferType"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "bankBrandName": {
            "type": "string"
          },
          "accountNumber": {
            "type": "string"
          },
          "transactionDate": {
            "type": "string"
          },
          "amountIn": {
            "type": "number"
          },
          "amountOut": {
            "type": "number"
          },
          "accumulated": {
            "type": "number",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "subCode": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string"
          },
          "transferType": {
            "type": "string",
            "enum": ["in", "out"]
          },
          "referenceCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string"
          }
        },
        "example": {
          "id": 123456789,
          "bankBrandName": "MB",
          "accountNumber": "56002005032008",
          "transactionDate": "2026-08-02 10:30:00",
          "amountIn": 500000,
          "amountOut": 0,
          "content": "ORDLKZM4A8X2",
          "transferType": "in"
        }
      },
      "WebhookResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": ["matched", "unmatched", "duplicate", "error"]
          },
          "message": {
            "type": "string"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "order_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "WebhookRetryStatusResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "pendingWebhooks": {
                "type": "integer"
              },
              "processing": {
                "type": "boolean"
              },
              "maxRetries": {
                "type": "integer"
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "nullable": true
          }
        }
      }
    }
  }
}